moved error handling into an 'errors.py' file
