+++ chime/__init__.py
@@ -8,8 +8,7 @@ logger = getLogger('chime')
 from os import mkdir
 from os.path import realpath, join
 
-from flask import Blueprint, Flask, current_app, render_template, session
-from .view_functions import common_template_args
+from flask import Blueprint, Flask
 
 from .httpd import run_apache_forever
 
@@ -106,12 +105,7 @@ def create_app(environ):
 
         logger.info("app config before_first_request: %s" % app.config)
 
-    @app.errorhandler(500)
-    def internal_server_error(error):
-        kwargs = common_template_args(current_app.config, session)
-        return render_template('error_500.html', **kwargs), 500
-
     return AppShim(app)
 
 # noinspection PyUnresolvedReferences
-from . import views
+from . import views, errors
